indexer/requirements.txt
dipdup==8.2.3
psycopg2-binary
pydantic
httpx
web3
